Move to scroll-to-top off the content
This moves the trigger for scroll-to-top out of the main content on the
FAQ page.
The trigger itself is slightly modified to have better hit area, to not
show shadow on hover (as the real shape is a sqaure while the content is an
icon within circle) and small visibility change on default state

Thanks!